data-prep/requirements.txt
# Data preparation tools requirements

# For PDF processing
pdf2image>=1.16.0
PyPDF2>=3.0.0

# For image generation
Pillow>=10.0.0

# For enhanced text extraction (optional)
# pdfplumber>=0.10.0
# pytesseract>=0.3.10  # For OCR if PDFs have images

# For future LLM-based tools
# openai>=1.0.0
# anthropic>=0.7.0
